+++ packages/dependency-extraction-webpack-plugin/lib/util.js
@@ -93,6 +93,7 @@ function defaultRequestToExternalModule( request ) {
 	switch ( request ) {
 		case '@wordpress/interactivity-router':
 		case '@wordpress/a11y':
+		case '@wordpress/polyfill':
 			return `import ${ request }`;
 	}